Stop declaring real files as phony
I've added some logic into Kati to attempt to prevent this type of
thing, since in the majority of cases, we expect a command to only run
if one of its dependencies has changed.

Workaround this specific case by declaring an output that is never
written. DO NOT use this elsewhere without permission. There will
eventually be a check for this workaround as well, but by that time
we'll probably be ready to mark rules that should not be cached.

Test: m RunStorageManagerRoboTests; m RunStorageManagerRoboTests
Change-Id: Ibfa6e5c275fda238849f773eb812994949fa8ab4
Merged-In: Ibfa6e5c275fda238849f773eb812994949fa8ab4