Remove unused robolectric prebuilt 4.8.2 directory
Due to the migration of all robolectric dependencies to a central
version located in /external/robolectric, we no longer need to keep the
prebuilt version in this directory. As mentioned by rexhoffman@, once
the dependency is eliminated from Wear, we are okay to delete these
files as they are now unused.

Bug: 321989200
Test: manual
Change-Id: I27cdbf72ce26b22f460766aadef63c4b980fb1af