Remove androidx annotation dependencies from media3 prebuilts
Updates the update-from-gmaven.py script to remove the dependencies on
androidx.annotation:annotation and
androidx.annotation:annotation-experimental from the media3 prebuilts.

The pom2bp rewrite rules for these dependencies are updated to rewrite
them to an empty string. A cleanup step is also added to the
fixAndroidBp function to remove the resulting empty strings from the
generated Android.bp file.

Bug: b/289815227
Test: Manually ran the script to generate new Android.bp
Change-Id: Iff93436f03931876740a9b59221f6550858987f0